Initializing readers and writers outside of the API request handlers
